feat(dapp): comments
