--- dapp/src/auctions/hooks/useAuctionDisplayImage.ts
@@ -1,22 +1,0 @@
-// Copyright (c) 2025 IOTA Stiftung
-// SPDX-License-Identifier: Apache-2.0
-
-import { useQuery } from '@tanstack/react-query';
-
-export function useAuctionDisplayImage(name: string, expirationTimestamp: string) {
-    return useQuery({
-        queryKey: ['generateSVG', name, expirationTimestamp],
-        queryFn: async () => {
-            const baseUrl = process.env.NEXT_PUBLIC_NAMES_DISPLAY_API_URL;
-            if (!baseUrl) {
-                throw new Error('NEXT_PUBLIC_NAMES_DISPLAY_API_URL is not configured');
-            }
-            const url = `${baseUrl}/${name}/${expirationTimestamp}`;
-            return {
-                url,
-            };
-        },
-        enabled: !!name && !!expirationTimestamp,
-        gcTime: 0,
-    });
-}
